+++ packages/renterd/lib/src/abstract/consensus_absts/accept_block_abst.dart
@@ -1,0 +1,10 @@
+import 'package:http/http.dart' as http;
+
+abstract class AcceptBlockAbst {
+  Future<http.Response> acceptBlock({
+    required String parentId,
+    required String address,
+    required String value,
+    required List<String> arbitraryDataList,
+  });
+}