+++ packages/renterd/lib/src/abstract/consensus_absts/get_sia_fund_fee_abst.dart
@@ -1,0 +1,5 @@
+import 'package:http/http.dart' as http;
+
+abstract class GetSiaFundFeeAbst {
+  Future<http.Response> getSiaFundFee({required int payout});
+}