+++ packages/renterd/lib/src/abstract/consensus_absts/get_state_abst.dart
@@ -1,0 +1,5 @@
+import 'package:http/http.dart' as http;
+
+abstract class GetStateAbst {
+  Future<http.Response> getState();
+}