+++ packages/renterd/lib/src/apis/consensus_api.dart
@@ -1,0 +1,3 @@
+const String postAcceptBlockApi = "api/bus/consensus/acceptblock";
+const String getStateApi = "api/bus/consensus/state";
+String getTotalPayoutApi(int payout) => "api/bus/consensus/siafundfee/$payout";