+++ packages/renterd/lib/src/data/services/accounts_services/get_an_account_by_id_impl.dart
@@ -7,11 +7,11 @@ import '../../../abstract/accounts_absts/get_an_account_by_id_abst.dart';
 import '../../../apis/account_api.dart';
 
 /// Project : Renterd's Package
-/// Description : This file contains the function  get An Account By Id whose Returns the account with the given ID. If it doesn't exist, it will be created. The provided host's key will be attached to the account as additional metadata., into response"
+/// Description : This file contains the function  [get An Account By Id] whose Returns the account with the given ID. If it doesn't exist, it will be created. The provided host's key will be attached to the account as additional metadata., into response"
 /// Author : James Brel
 /// onCreated : 14/05/2023
 /// onUpadted : #
-/// Upadted by : James Brel
+/// Upadted by : #
 class GetAnAccountByIdImpl implements GetAnAccountByIdAbst {
   @override
   Future<http.Response> getAnAccountById({