+++ packages/renterd/lib/src/data/services/accounts_services/lock_an_account_impl.dart
@@ -7,11 +7,11 @@ import '../../../abstract/accounts_absts/lock_an_account_abst.dart';
 import '../../../apis/account_api.dart';
 
 /// Project : Renterd's Package
-/// Description : This file contains the function Lock An Account whose Commonly used by the worker, this endpoint allows for locking an account either exclusively or not. into response"
+/// Description : This file contains the function [Lock An Account] whose Commonly used by the worker, this endpoint allows for locking an account either exclusively or not. into response"
 /// Author : James Brel
 /// onCreated : 15/05/2023
 /// onUpadted : #
-/// Upadted by : James Brel
+/// Upadted by : #
 class LockAnAccountImpl implements LockAnAccountAbst {
   @override
   Future<http.Response> lockAnAccount({