+++ packages/renterd/CHANGELOG.md
@@ -1,7 +1,11 @@
-
+## 0.0.2
+### Consensus Services Created
+  - [x] Accept Block
+  - [x] Get Sia Fund Fee
+  - [x] Get State
+---------------------------------
 
 ## 0.0.1
-
 ### Accounts Services Created
   - [x] Add Deposit
   - [x] Get All Accounts